# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentación automática del backend de Taskly",
        "title": "Taskly API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Autentica al usuario con email y contraseña. Retorna un token JWT válido por 24 horas para usar en endpoints protegidos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión de usuario",
                "parameters": [
                    {
                        "description": "Credenciales de acceso",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "schema": {
                            "$ref": "#/definitions/models.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado o contraseña incorrecta",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna los datos públicos del usuario que inició sesión, obtenidos desde el token JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener perfil del usuario autenticado",
                "responses": {
                    "200": {
                        "description": "Perfil obtenido exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite modificar el nombre y/o contraseña del usuario logueado. Solo los campos enviados serán actualizados.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar perfil del usuario autenticado",
                "parameters": [
                    {
                        "description": "Campos a actualizar",
                        "name": "datos",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Perfil actualizado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.UserProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o error al encriptar",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/registro": {
            "post": {
                "description": "Crea una cuenta de usuario con email y contraseña. La contraseña se encripta automáticamente con bcrypt antes de guardar.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar un nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "usuario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario registrado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Usuario"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o error al encriptar contraseña",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/usuarios/buscar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite buscar un usuario existente usando su email. Útil para verificar disponibilidad antes de registrar.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Buscar usuario por email",
                "parameters": [
                    {
                        "type": "string",
                        "example": "ana@example.com",
                        "description": "Email del usuario a buscar",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.UserProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Parámetro email faltante",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Equipo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "miembros": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Miembro"
                    }
                },
                "nombre": {
                    "type": "string"
                },
                "ownerid": {
                    "type": "integer"
                },
                "tareas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tarea"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Datos inválidos"
                },
                "message": {
                    "type": "string",
                    "example": "Campo requerido"
                }
            }
        },
        "models.LoginRequest": {
            "description": "Credenciales requeridas para autenticación",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ana@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "MiClave123"
                }
            }
        },
        "models.LoginResponse": {
            "description": "Contiene el token JWT para autenticar requests posteriores",
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
        "models.Miembro": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "equipo": {
                    "$ref": "#/definitions/models.Equipo"
                },
                "equipoid": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "rol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usuario": {
                    "$ref": "#/definitions/models.Usuario"
                },
                "usuarioid": {
                    "type": "integer"
                }
            }
        },
        "models.RegisterRequest": {
            "description": "Información requerida para crear una cuenta",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ana@example.com"
                },
                "nombre": {
                    "type": "string",
                    "example": "Ana López"
                },
                "password": {
                    "type": "string",
                    "example": "MiClave123"
                }
            }
        },
        "models.Tarea": {
            "type": "object",
            "properties": {
                "asignadoid": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "descripcion": {
                    "type": "string"
                },
                "equipoid": {
                    "type": "integer"
                },
                "estado": {
                    "type": "string"
                },
                "fechaLimite": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "prioridad": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateProfileRequest": {
            "description": "Campos opcionales para actualizar. Solo los enviados serán modificados.",
            "type": "object",
            "properties": {
                "nombre": {
                    "type": "string",
                    "example": "Ana López Actualizado"
                },
                "password": {
                    "type": "string",
                    "example": "NuevaClave456"
                }
            }
        },
        "models.UserProfileResponse": {
            "description": "Datos visibles del usuario autenticado",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ana@example.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "nombre": {
                    "type": "string",
                    "example": "Ana López"
                }
            }
        },
        "models.Usuario": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "equipos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipo"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}